+++ src/server/admin/admin-profile/admin-profile.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import { SequelizeModule } from '@nestjs/sequelize';
+import { AdministratorModel } from 'share/models/administrator.model';
+import { AdminProfileController } from './admin-profile.controller';
+import { AdminProfileService } from './admin-profile.service';
+
+@Module({
+  controllers: [AdminProfileController],
+  imports: [SequelizeModule.forFeature([AdministratorModel])],
+  providers: [AdminProfileService],
+})
+export class AdminProfileModule {}
