+++ src/server/app.service.ts
@@ -1,11 +1,8 @@
 import { Injectable } from '@nestjs/common';
-import { AdministratorRepository } from 'app.repository';
 import { sequelize } from 'vendors/sequelize/sequelize';
 
 @Injectable()
 export class AppService {
-  constructor(private repository: AdministratorRepository) {}
-
   getHello(): string {
     return 'Hello World!';
   }
@@ -16,10 +13,5 @@ export class AppService {
       NODE_ENV: process.env.NODE_ENV,
       DB: 'ok',
     };
-  }
-
-  async test() {
-    const count = await this.repository.count();
-    return `count: ${count}`;
   }
 }
